backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6
redis>=4.5.5

# LLM and AI dependencies
openai>=1.0.0
langchain>=0.0.267
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.2

# Vector database
chromadb>=0.4.6
pinecone-client>=2.2.1
weaviate-client>=3.15.4

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0
httpx>=0.24.0

# Utilities
tqdm>=4.65.0
requests>=2.30.0
aiohttp>=3.8.4
websockets>=11.0.3
pyyaml>=6.0
jinja2>=3.1.2
